refactor: delete the phantom ports — dead web/hybrid hexagon, services/ports fiction, empty ports/ package
EvaluationsRepository/DimensionsRepository had 3 adapters each and zero
production callers (the shipped shared-evaluations feature reads via the
git-clone mirror, not web repositories). services/ports.py claimed to be
the single services->data crossing while 28 of 31 crossings bypassed it;
service code now imports data.fs.report_parser directly and the API goes
through the honest services/run_reports facade. The RunStorage Protocol
(no implementation, no callers) and the docstring-only ports/ package are
gone. The two _ensure_fresh private reaches now use ensure_projected().
