+++ src/quodeq/api/routes_rescore.py
@@ -7,7 +7,7 @@ from http import HTTPStatus
 from flask import Flask, Response, jsonify, request
 
 from quodeq.api.helpers import error_response
-from quodeq.services.ports import list_runs, read_run_data
+from quodeq.services.run_reports import list_runs, read_run_data
 from quodeq.services.deleted import deleted_keys as load_deleted_keys
 from quodeq.services.dismissed import dismissed_keys as load_dismissed_keys
 from quodeq.services.rescore import rescore_dimensions
