--- src/quodeq/ports/__init__.py
@@ -1,1 +1,0 @@
-"""Documentation-only namespace. Real ports live elsewhere; see CONTEXT.md."""
