+++ src/quodeq/services/_fs_metadata.py
@@ -8,7 +8,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 from quodeq.data.fs.standards_prefs import load_visible_standard_ids
-from quodeq.services.ports import RunInfo, read_run_data, safe_read_dir, summarize_dimensions
+from quodeq.data.fs.report_parser.grades import summarize_dimensions
+from quodeq.data.fs.report_parser.runs import RunInfo, read_run_data, safe_read_dir
 from quodeq.shared.validation import validate_path_segment
 
 _logger = logging.getLogger(__name__)
