+++ src/quodeq/services/_runs_unit.py
@@ -9,7 +9,8 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from quodeq.services.ports import most_frequent_grade, parse_numeric_score, read_run_scalars
+from quodeq.data.fs.report_parser.grades import most_frequent_grade, parse_numeric_score
+from quodeq.data.fs.report_parser.runs import read_run_scalars
 from quodeq.data.sqlite.run_index import (
     RunRow,
     list_runs_for_project,
