--- tests/services/test_ports_reexports.py
@@ -1,5 +1,0 @@
-"""Verify the services layer re-exports new data-layer protocols."""
-
-
-def test_findings_repository_re_exported():
-    from quodeq.services.ports import FindingsRepository  # noqa: F401
